requirements.txt
# Core scientific computing
numpy>=1.24.0
scipy>=1.10.0
pandas>=2.0.0
astropy>=5.3.0

# Machine learning & AI
scikit-learn>=1.3.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.15.0

# Configuration & utilities
pyyaml>=6.0
python-dotenv>=1.0.0
click>=8.1.0
rich>=13.0.0
tqdm>=4.66.0
python-dateutil>=2.8.2

# Data handling
h5py>=3.8.0
xarray>=2023.1.0
